JSON value access error in submission and editing methods
Multiple methods related to submission or editing had similar errors when attempting to access a returned JSON object.

Many tried to get a root node named "json" from a json object. No such node existed (or was already parsed out in the WebAgent's ExecuteRequestAsync method) and so an exception was thrown.

Many incorrectly parsed any returned ratelimit error (the "You are doing that too much." error).

Related to the ratelimit error, a few of the methods incorrectly checked if the json object returned any errors.